+++ src/map.h
@@ -30,8 +30,8 @@
  */
 void find_search_intervals(Synmap * syn, FILE * intfile, bool pblock);
 
-float calculate_score(long start, long stop, Block * blk);
+double calculate_score(long start, long stop, Block * blk);
 
-float calculate_target_score(long a1, long a2, Block * bounds[2]);
+double calculate_target_score(long a1, long a2, Block * bounds[2]);
 
 #endif
